Create HomeScreen class. Remove some code in main.dart.
